# clients/java/api/openapi.yaml
openapi: 3.1.0
info:
  description: "Define, evaluate, and enforce data retention policies"
  title: DRPE — Data Retention Policy Engine
  version: 0.1.0
servers:
- description: Local DRPE API
  url: http://localhost:8000
paths:
  /api/v1/policies/validate:
    post:
      operationId: validate_policy_api_v1_policies_validate_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ValidateRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidateResponse"
          description: Successful Response
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
          description: Validation Error
      security:
      - BearerAuth: []
      summary: Validate Policy
      tags:
      - policies
      x-content-type: application/json
      x-accepts:
      - application/json
  /api/v1/policies/import:
    post:
      operationId: import_policies_api_v1_policies_import_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ImportRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportResponse"
          description: Successful Response
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
          description: Validation Error
      security:
      - BearerAuth: []
      summary: Import Policies
      tags:
      - policies
      x-content-type: application/json
      x-accepts:
      - application/json
  /api/v1/policies:
    get:
      operationId: list_policies_api_v1_policies_get
      parameters:
      - explode: true
        in: query
        name: status_filter
        required: false
        schema:
          nullable: true
          type: string
        style: form
      - explode: true
        in: query
        name: policy_kind
        required: false
        schema:
          $ref: "#/components/schemas/PolicyKind"
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/PolicyListItem"
                type: array
          description: Successful Response
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
          description: Validation Error
      security:
      - BearerAuth: []
      summary: List Policies
      tags:
      - policies
      x-accepts:
      - application/json
    post:
      operationId: create_policy_api_v1_policies_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PolicyCreateRequest"
        required: true
      responses:
        "201":
          content:
            application/json:
              schema: {}
          description: Successful Response
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
          description: Validation Error
      security:
      - BearerAuth: []
      summary: Create Policy
      tags:
      - policies
      x-content-type: application/json
      x-accepts:
      - application/json
  /api/v1/policies/{policy_id}:
    delete:
      operationId: delete_policy_api_v1_policies__policy_id__delete
      parameters:
      - explode: false
        in: path
        name: policy_id
        required: true
        schema:
          title: Policy Id
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema: {}
          description: Successful Response
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
          description: Validation Error
      security:
      - BearerAuth: []
      summary: Delete Policy
      tags:
      - policies
      x-accepts:
      - application/json
    get:
      operationId: get_policy_api_v1_policies__policy_id__get
      parameters:
      - explode: false
        in: path
        name: policy_id
        required: true
        schema:
          title: Policy Id
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema: {}
          description: Successful Response
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
          description: Validation Error
      security:
      - BearerAuth: []
      summary: Get Policy
      tags:
      - policies
      x-accepts:
      - application/json
    put:
      operationId: update_policy_api_v1_policies__policy_id__put
      parameters:
      - explode: false
        in: path
        name: policy_id
        required: true
        schema:
          title: Policy Id
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PolicyCreateRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema: {}
          description: Successful Response
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
          description: Validation Error
      security:
      - BearerAuth: []
      summary: Update Policy
      tags:
      - policies
      x-content-type: application/json
      x-accepts:
      - application/json
  /api/v1/policies/{policy_id}/status:
    post:
      operationId: change_policy_status_api_v1_policies__policy_id__status_post
      parameters:
      - explode: false
        in: path
        name: policy_id
        required: true
        schema:
          title: Policy Id
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PolicyStatusChangeRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema: {}
          description: Successful Response
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
          description: Validation Error
      security:
      - BearerAuth: []
      summary: Change Policy Status
      tags:
      - policies
      x-content-type: application/json
      x-accepts:
      - application/json
  /api/v1/policies/{policy_id}/versions:
    get:
      operationId: list_policy_versions_api_v1_policies__policy_id__versions_get
      parameters:
      - explode: false
        in: path
        name: policy_id
        required: true
        schema:
          title: Policy Id
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/PolicyVersionInfo"
                type: array
          description: Successful Response
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
          description: Validation Error
      security:
      - BearerAuth: []
      summary: List Policy Versions
      tags:
      - policies
      x-accepts:
      - application/json
  /api/v1/policies/{policy_id}/versions/{ver}:
    get:
      operationId: get_policy_version_api_v1_policies__policy_id__versions__ver__get
      parameters:
      - explode: false
        in: path
        name: policy_id
        required: true
        schema:
          title: Policy Id
          type: string
        style: simple
      - explode: false
        in: path
        name: ver
        required: true
        schema:
          title: Ver
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema: {}
          description: Successful Response
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
          description: Validation Error
      security:
      - BearerAuth: []
      summary: Get Policy Version
      tags:
      - policies
      x-accepts:
      - application/json
  /api/v1/policies/{policy_id}/versions/{ver}/activate:
    post:
      operationId: activate_policy_version_api_v1_policies__policy_id__versions__ver__activate_post
      parameters:
      - explode: false
        in: path
        name: policy_id
        required: true
        schema:
          title: Policy Id
          type: string
        style: simple
      - explode: false
        in: path
        name: ver
        required: true
        schema:
          title: Ver
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema: {}
          description: Successful Response
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
          description: Validation Error
      security:
      - BearerAuth: []
      summary: Activate Policy Version
      tags:
      - policies
      x-accepts:
      - application/json
  /api/v1/policies/{policy_id}/diff:
    post:
      operationId: diff_policy_versions_api_v1_policies__policy_id__diff_post
      parameters:
      - explode: false
        in: path
        name: policy_id
        required: true
        schema:
          title: Policy Id
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PolicyDiffRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PolicyDiffResponse"
          description: Successful Response
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
          description: Validation Error
      security:
      - BearerAuth: []
      summary: Diff Policy Versions
      tags:
      - policies
      x-content-type: application/json
      x-accepts:
      - application/json
...
components:
  schemas:
...
    HTTPValidationError:
      example:
        detail:
        - msg: msg
          loc:
          - ValidationError_loc_inner
          - ValidationError_loc_inner
          input: ""
          ctx: "{}"
          type: type
        - msg: msg
          loc:
          - ValidationError_loc_inner
          - ValidationError_loc_inner
          input: ""
          ctx: "{}"
          type: type
      properties:
        detail:
          items:
            $ref: "#/components/schemas/ValidationError"
          type: array
      title: HTTPValidationError
...
    ImportRequest:
      example:
        yaml: yaml
      properties:
        yaml:
          title: Yaml
          type: string
      required:
      - yaml
      title: ImportRequest
    ImportResponse:
      example:
        imported:
        - imported
        - imported
        count: 0
      properties:
        imported:
          items:
            type: string
          type: array
        count:
          title: Count
          type: integer
      required:
      - count
      - imported
      title: ImportResponse
...
    PolicyCreateRequest:
      description: Create from YAML string or JSON policy object.
      example:
        yaml: yaml
        policy:
          key: ""
      properties:
        yaml:
          nullable: true
          type: string
        policy:
          additionalProperties: true
          nullable: true
          type: object
      title: PolicyCreateRequest
    PolicyDiffChange:
      description: A single structural change between two policy versions.
      example:
        path: path
        op: add
        new: ""
        old: ""
      properties:
        path:
          title: Path
          type: string
        op:
          enum:
          - add
          - remove
          - replace
          title: Op
          type: string
        old: {}
        new: {}
      required:
      - op
      - path
      title: PolicyDiffChange
    PolicyDiffRequest:
      example:
        from_version: 1
        to_version: 1
      properties:
        from_version:
          minimum: 1.0
          title: From Version
          type: integer
        to_version:
          minimum: 1.0
          title: To Version
          type: integer
      required:
      - from_version
      - to_version
      title: PolicyDiffRequest
    PolicyDiffResponse:
      example:
        policy_id: policy_id
        changes:
        - path: path
          op: add
          new: ""
          old: ""
        - path: path
          op: add
          new: ""
          old: ""
        from_version: 0
        to_version: 6
      properties:
        policy_id:
          title: Policy Id
          type: string
        from_version:
          title: From Version
          type: integer
        to_version:
          title: To Version
          type: integer
        changes:
          items:
            $ref: "#/components/schemas/PolicyDiffChange"
          type: array
      required:
      - from_version
      - policy_id
      - to_version
      title: PolicyDiffResponse
    PolicyKind:
      enum:
      - retention
      - classification
      title: PolicyKind
      type: string
    PolicyListItem:
      example:
        entity_count: 6
        scope_data_types:
        - scope_data_types
        - scope_data_types
        jurisdiction: jurisdiction
        scope_sources:
        - scope_sources
        - scope_sources
        version: 0
        policy_kind: retention
        data_classification: data_classification
        excluded_data_types:
        - excluded_data_types
        - excluded_data_types
        name: name
        excluded_sources:
        - excluded_sources
        - excluded_sources
        id: id
        status: draft
        rule_count: 1
      properties:
        id:
          title: Id
          type: string
        name:
          title: Name
          type: string
        version:
          title: Version
          type: integer
        status:
          $ref: "#/components/schemas/PolicyStatus"
        jurisdiction:
          title: Jurisdiction
          type: string
        policy_kind:
          $ref: "#/components/schemas/PolicyKind"
        data_classification:
          nullable: true
          type: string
        entity_count:
          nullable: true
          type: integer
        scope_data_types:
          items:
            type: string
          type: array
        scope_sources:
          items:
            type: string
          type: array
        excluded_data_types:
          items:
            type: string
          type: array
        excluded_sources:
          items:
            type: string
          type: array
        rule_count:
          title: Rule Count
          type: integer
      required:
      - id
      - jurisdiction
      - name
      - policy_kind
      - rule_count
      - status
      - version
      title: PolicyListItem
...
    PolicyStatus:
      enum:
      - draft
      - active
      - deprecated
      - archived
      title: PolicyStatus
      type: string
    PolicyStatusChangeRequest:
      example:
        status: draft
      properties:
        status:
          $ref: "#/components/schemas/PolicyStatus"
      required:
      - status
      title: PolicyStatusChangeRequest
    PolicyVersionInfo:
      description: Lightweight metadata for a stored policy version snapshot.
      example:
        policy_id: policy_id
        name: name
        created_at: 2000-01-23T04:56:07.000+00:00
        version: 0
        status: status
      properties:
        policy_id:
          title: Policy Id
          type: string
        version:
          title: Version
          type: integer
        created_at:
          format: date-time
          nullable: true
          type: string
        name:
          nullable: true
          type: string
        status:
          nullable: true
          type: string
      required:
      - policy_id
      - version
      title: PolicyVersionInfo
...
    ValidateRequest:
      example:
        yaml: yaml
        policy:
          key: ""
      properties:
        yaml:
          nullable: true
          type: string
        policy:
          additionalProperties: true
          nullable: true
          type: object
      title: ValidateRequest
    ValidateResponse:
      example:
        valid: true
        policy_kind: null
        classification_policy:
          owner: owner
          jurisdiction: jurisdiction
          text_fields:
          - text_fields
          - text_fields
          rules:
          - condition:
              all:
              - field: field
                value: ""
                operator: eq
              - field: field
                value: ""
                operator: eq
              any:
              - field: field
                value: ""
                operator: eq
              - field: field
                value: ""
                operator: eq
            description: description
            action: flag
            handling: handling
            id: id
            priority: 1
          - condition:
              all:
              - field: field
                value: ""
                operator: eq
              - field: field
                value: ""
                operator: eq
              any:
              - field: field
                value: ""
                operator: eq
              - field: field
                value: ""
                operator: eq
            description: description
            action: flag
            handling: handling
            id: id
            priority: 1
          version: 1
          tags:
          - tags
          - tags
          policy_kind: null
          expires_at: null
          entities:
          - detection:
              catalog_ref: catalog_ref
              regex: regex
              field_names:
              - field_names
              - field_names
              ner_types:
              - ner_types
              - ner_types
            regulatory_refs:
            - regulatory_refs
            - regulatory_refs
            id: id
            label: label
            sensitivity: low
            classification: null
          - detection:
              catalog_ref: catalog_ref
              regex: regex
              field_names:
              - field_names
              - field_names
              ner_types:
              - ner_types
              - ner_types
            regulatory_refs:
            - regulatory_refs
            - regulatory_refs
            id: id
            label: label
            sensitivity: low
            classification: null
          scope:
            sources:
            - sources
            - sources
            exclude:
              sources:
              - sources
              - sources
              data_types:
              - data_types
              - data_types
            data_types:
            - data_types
            - data_types
          name: name
          id: id
          effective_from: null
          status: null
        errors:
        - errors
        - errors
        policy:
          owner: owner
          jurisdiction: jurisdiction
          rules:
          - requires_approval: false
            retain_until: retain_until
            condition:
              all:
              - field: field
                value: ""
                operator: eq
              - field: field
                value: ""
                operator: eq
              any:
              - field: field
                value: ""
                operator: eq
              - field: field
                value: ""
                operator: eq
            archive_target: archive_target
            notify_before: notify_before
            description: description
            action: retain
            id: id
            priority: 1
            grace_period: grace_period
          - requires_approval: false
            retain_until: retain_until
            condition:
              all:
              - field: field
                value: ""
                operator: eq
              - field: field
                value: ""
                operator: eq
              any:
              - field: field
                value: ""
                operator: eq
              - field: field
                value: ""
                operator: eq
            archive_target: archive_target
            notify_before: notify_before
            description: description
            action: retain
            id: id
            priority: 1
            grace_period: grace_period
          version: 0
          tags:
          - tags
          - tags
          dsar:
            erasure_exceptions:
            - erasure_exceptions
            - erasure_exceptions
            response_deadline: response_deadline
            right_to_access: true
            right_to_erasure: true
          policy_kind: retention
          data_classification: PII
          expires_at: 2000-01-23
          audit:
            log_evaluations: true
            retention_of_audit_logs: retention_of_audit_logs
            log_actions: true
          scope:
            sources:
            - sources
            - sources
            exclude:
              sources:
              - sources
              - sources
              data_types:
              - data_types
              - data_types
            data_types:
            - data_types
            - data_types
          name: name
          id: id
          effective_from: 2000-01-23
          status: draft
      properties:
        valid:
          title: Valid
          type: boolean
        policy:
          $ref: "#/components/schemas/Policy"
        classification_policy:
          $ref: "#/components/schemas/ClassificationPolicy"
        policy_kind:
          $ref: "#/components/schemas/PolicyKind"
        errors:
          items:
            type: string
          type: array
      required:
      - valid
      title: ValidateResponse
    ValidationError:
      example:
        msg: msg
        loc:
        - ValidationError_loc_inner
        - ValidationError_loc_inner
        input: ""
        ctx: "{}"
        type: type
      properties:
        loc:
          items:
            $ref: "#/components/schemas/ValidationError_loc_inner"
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
        input: {}
        ctx:
          title: Context
          type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
    ValidationError_loc_inner:
      anyOf:
      - type: string
      - type: integer
  securitySchemes:
    BearerAuth:
      description: API key as Bearer token (DRPE_API_KEY)
      scheme: bearer
      type: http